Desktop/Smart Lines/DjangoSaveJune13/djangoWorks/restAPI/djangorest/requirement.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-cors-headers==2.3.0
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
PyGithub==1.35
PyJWT==1.6.1
python-decouple==3.1
pytz==2018.4
requests==2.18.4
selenium==3.12.0
urllib3==1.22
